requirements.txt
streamlit
google-generativeai
python-docx
Pillow
streamlit-gsheets
pandas